+++ host/index.jsx
@@ -11,4 +11,5 @@
 #include "modules/misc.jsx"
 #include "modules/bubbletext.jsx"
 #include "modules/aetoam.jsx"
+#include "modules/controller.jsx"
 #include "modules/api.jsx"