+++ host/modules/api.jsx
@@ -5,5 +5,15 @@ $.global.FishTools = {
     readVelocity: _readVelocity,
     applyVelocity: _applyVelocity,
     applyElastic: _applyElastic,
-    debugProps: _debugProps
+    debugProps: _debugProps,
+    controllerRead: _controllerRead,
+    controllerSetValue: _controllerSetValue,
+    controllerSetAxis: _controllerSetAxis,
+    controllerSetBlendMode: _controllerSetBlendMode,
+    controllerSetVector: _controllerSetVector,
+    controllerSetPositionAxis: _controllerSetPositionAxis,
+    controllerSetPositionXY: _controllerSetPositionXY,
+    controllerSetPositionZ: _controllerSetPositionZ,
+    controllerBeginUndo: _controllerBeginUndo,
+    controllerEndUndo: _controllerEndUndo
 };